pyproject.toml
[project]
name = "ray-convergence-engine"
version = "0.1.0"
description = "Data State-First convergence engine using Ray + Lance"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ray[serve]>=2.40.0",
    "starlette>=0.36.3",
    "requests>=2.31.0",
    "pylance>=0.10.15",
    "pyarrow>=14.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
